flask/requirements.txt
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cssselect2==0.4.1
defusedxml==0.7.1
drawSvg==1.8.2
Flask==2.0.1
idna==3.2
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
Pillow==8.3.2
pycparser==2.20
requests==2.26.0
tinycss2==1.1.0
urllib3==1.26.6
webencodings==0.5.1
Werkzeug==2.0.1
gunicorn==20.1.0